LayoutTests/fast/frames/adopt-object-into-itself-expected.txt

ALERT: HierarchyRequestError: The operation would yield an incorrect node tree.
